requirements.txt
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.9
python-dotenv==1.0.0
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
